Makefile.uk: Change `addlib` to `addgoapp`
Golang applications which use the `libgo` external library have
custom build mechanics, where the `addlib` rule is replaced by
either `addgolib` (for GO packages used as libraries), or `addgoapp`.
In this case, we need to call the `addgoapp` Makefile rule to
indicate that we're building a GO app which provides the main function.

Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>
Reviewed-by: Radu Nichita <radunichita99@gmail.com>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #9